Seperate class into seperate files
