+++ src/Microsoft.DocAsCode.EntityModel/ExtractMetadata/SupportedSourceType.cs
@@ -1,0 +1,14 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace Microsoft.DocAsCode.EntityModel.ExtractMetadata
+{
+    internal enum SupportedSourceType
+    {
+        VBSourceCode,
+        VBProject,
+        CSharpSourceCode,
+        CSharpProject,
+        Solution
+    }
+}
